Move LoadPrivatekey() to utils package
