+++ utils/utils.go
@@ -1,8 +1,10 @@
 package utils
 
 import (
-	"io/ioutil"
 	"crypto/rsa"
+	"fmt"
+	"io/ioutil"
+
 	"github.com/dgrijalva/jwt-go"
 )
 
@@ -14,4 +16,16 @@ func ReadPublicKey() (*rsa.PublicKey, error) {
 	}
 	ParsedKey, err := jwt.ParseRSAPublicKeyFromPEM(Key)
 	return ParsedKey, err
+}
+
+func LoadPrivateKey() (*rsa.PrivateKey, error) {
+	Key, err := ioutil.ReadFile("private-key.pem")
+	if err != nil {
+		return nil, fmt.Errorf("failed to road private key: %s", err)
+	}
+	PrivateKey, err := jwt.ParseRSAPrivateKeyFromPEM(Key)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse Privatekey: %s", err)
+	}
+	return PrivateKey, nil
 }
